fixed header full screen only shows first time, created styles folder with all custom styles
