+++ src/main.tsx
@@ -1,9 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
-import "./App.css";
-import "./index.css";
-import "./carousel.css";
+import "./styles/App.css";
+import "./styles/index.css";
+import "./styles/carousel.css";
+import "./styles/headerAnimation.css";
 import { BrowserRouter } from "react-router-dom";
 import { MyContextProvider } from "./constants/context.tsx";
