+++ src/styles/carousel.css
